Ocean/QAP/requirements.txt
numpy==1.24.3
pandas==2.0.1
dimod==0.12.5
dwave-neal==0.6.0
dwave-tabu==0.5.0
matplotlib==3.7.1